refactor(runtime): move ICustomAttribute+ and ICustomElement+ interfaces
