+++ packages/runtime/src/aurelia.ts
@@ -1,6 +1,7 @@
 import { DI, IContainer, IRegistry, PLATFORM, Registration } from '@aurelia/kernel';
 import { LifecycleFlags } from './observation';
-import { ICustomElement, IRenderingEngine } from './templating/lifecycle-render';
+import { ICustomElement } from './templating/custom-element';
+import { IRenderingEngine } from './templating/lifecycle-render';
 
 export interface ISinglePageApp {
   host: any;
