+++ packages/runtime/src/definitions.ts
@@ -1,11 +1,12 @@
 // tslint:disable:no-reserved-keywords
-import { Constructable, DI, Immutable, Omit, PLATFORM } from '@aurelia/kernel';
+import { DI, Immutable, Omit, PLATFORM } from '@aurelia/kernel';
 import { ForOfStatement, Interpolation, IsBindingBehavior } from './binding/ast';
 import { BindingMode } from './binding/binding-mode';
 import { DelegationStrategy } from './binding/event-manager';
 import { INode } from './dom';
 import { IResourceDefinition, ResourceDescription } from './resource';
-import { ICustomElement, ICustomElementHost } from './templating/lifecycle-render';
+import { CustomElementConstructor, ICustomElement } from './templating/custom-element';
+import { ICustomElementHost } from './templating/lifecycle-render';
 
 /*@internal*/
 export const customElementName = 'custom-element';
@@ -211,12 +212,6 @@ export interface ILetBindingInstruction extends ITargetedInstruction {
   from: string | IsBindingBehavior | Interpolation;
   to: string;
 }
-
-type CustomElementStaticProperties = Pick<TemplateDefinition, 'containerless' | 'shadowOptions' | 'bindables'>;
-type CustomAttributeStaticProperties = Pick<AttributeDefinition, 'bindables'>;
-
-export type CustomElementConstructor = Constructable & CustomElementStaticProperties;
-export type CustomAttributeConstructor = Constructable & CustomAttributeStaticProperties;
 
 /*@internal*/
 export const buildRequired: IBuildInstruction = Object.freeze({
