+++ packages/runtime/src/html-renderer.ts
@@ -1,4 +1,4 @@
-import { inject, IRegistry, IContainer } from '@aurelia/kernel';
+import { IContainer, inject, IRegistry } from '@aurelia/kernel';
 import { Binding } from './binding/binding';
 import { BindingMode } from './binding/binding-mode';
 import { Call } from './binding/call';
@@ -12,7 +12,9 @@ import { Ref } from './binding/ref';
 import { customAttributeKey, customElementKey, ICallBindingInstruction, IHydrateAttributeInstruction, IHydrateElementInstruction, IHydrateTemplateController, IInterpolationInstruction, IIteratorBindingInstruction, ILetElementInstruction, IListenerBindingInstruction, IPropertyBindingInstruction, IRefBindingInstruction, ISetAttributeInstruction, ISetPropertyInstruction, IStylePropertyBindingInstruction, ITextBindingInstruction, TargetedInstructionType, TemplatePartDefinitions } from './definitions';
 import { DOM, INode, IRemovableNode } from './dom';
 import { IAttach, IAttachables, IBindables, IBindScope, IRenderable, IRenderContext } from './lifecycle';
-import { ICustomAttribute, ICustomElement, IInstructionRenderer, instructionRenderer, IRenderer, IRenderingEngine } from './templating/lifecycle-render';
+import { ICustomAttribute } from './templating/custom-attribute';
+import { ICustomElement } from './templating/custom-element';
+import { IInstructionRenderer, instructionRenderer, IRenderer, IRenderingEngine } from './templating/lifecycle-render';
 
 export function ensureExpression<TFrom>(parser: IExpressionParser, srcOrExpr: TFrom, bindingType: BindingType): Exclude<TFrom, string> {
   if (typeof srcOrExpr === 'string') {
